+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.0.7 (unreleased)
 ------------------
 
+- Hide unwanted upgrades from site-creation and quickinstaller
+  [boulch]
+
 - Move local manager role and sharing permissions to imio.smartweb.common
   Use new common.interfaces.ILocalManagerAware to mark a locally manageable content
   [boulch]